move the authenticator test into own file
