+++ src/client/mod.rs
@@ -10,6 +10,8 @@ pub mod params_test;
 pub mod storage_test;
 #[cfg(test)]
 pub mod mock_client_test;
+#[cfg(test)]
+pub mod authenticator_test;
 
 use futures::future::FutureResult;
 use errors::Error;
